tests/submission/Opera/getElementsByClassName/007.htm

<!doctype html>
<html>
 <head>
  <title>document.getElementsByClassName(): multiple classes</title>
  <script src="../resources/testharness.js"></script>
 </head>
 <body class="a b">
  <div id="log"></div>
  <script> test(function() {
                  assert_array_equals(document.getElementsByClassName("b\t\f\n\na\rb"), [document.body]);
               })
  </script>
 </body>
</html>
